structurize rendering content
