+++ Sources/AlgoliaSearchClient/Models/Search/Response/SearchResponse/Auxiliary/RenderingContent/FacetOrderContainer.swift
@@ -1,0 +1,18 @@
+//
+//  FacetOrderContainer.swift
+//  
+//
+//  Created by Vladislav Fitc on 17/03/2021.
+//
+
+import Foundation
+
+public struct FacetOrderContainer: Codable {
+      
+  public let facetOrder: [AttributedFacets]
+  
+  public init(facetOrder: [AttributedFacets] = []) {
+    self.facetOrder = facetOrder
+  }
+  
+}
