+++ Sources/AlgoliaSearchClient/Models/Search/Response/SearchResponse/Auxiliary/RenderingContent/Redirect.swift
@@ -1,0 +1,16 @@
+//
+//  Redirect.swift
+//  
+//
+//  Created by Vladislav Fitc on 17/03/2021.
+//
+
+import Foundation
+
+/// Container for Redirect information
+public struct Redirect: Codable {
+  
+  /// Redirect URL
+  public let url: URL?
+  
+}
